bin/memo: don't fail because the outpipe does
When memo reads from a cache and the program after memo sends a SIGPIPE back
(for whatever reason), memo would interpret that as a failed attempt to find
the cached file, therefore trying to recreate the cache (which would also
not work as the output pipe was already closed). Double fail.