bin/memo: add slurp stdin option (-s)
Also redesign the script so it can deal with using the stdin as part of the
hashed content better.

I'm dissatisfied with some parts, like the duplication of the (de)compressor
logic, but I'll have to wait until I get some bright idea to make this
elegant. As for now, the functionality is king. I'm already loving the hell
out of the -s switch.