bin/memo: make -s (use stdin) flag work on macOS
macOS does not have procfs, thus no /proc/<pid>/fd/<fd> (funny how I had
mentioned portability in the code...). Create shell-specific (zsh and
bash-compatibles only) ways for cleaning up, even when the script is
sourced instead of executed in a standalone fashion.

It would've been nicer to use the `-v` test for variable existence, but
it was only introduced in bash 4.2.

Tested on macOS.