vim: fix [-*+] formatlistpat for all fts
This used to be copied from the markdown filetype. There we could live
without matching whitespace at the start because it uses:

  set comments=b:-,b:*,b:+

Meaning Vim recognizes the list starters as comments too. Which means
they're implicitly taken into account for indentation. For other
languages, where comments is something else, we must match the
whitespace in the formatlistpat match.

The other change in this CL is to match more numbered list patterns.
E.g.:

    1]
    2]

Or

    1)
    2)

Or even

    1:
    2:

Though I'm not sure I will keep that. It could be annoying as it occurs
in normal text flow too.