bin/memo: shellcheck: separate assign/declare
Split up assignment and declaration so return codes aren't lost. In this
case, the return code was mostly '1' (error), because

     (( opt_stdin )) && tee >(compressor > "$stdinsave")

Mostly yields false (opt_stdin == 0).

I'm not sure whether tee(1) exits with an error code if the input stream
is interrupted somehow. That would be nice, as we don't want to save
that hash then anyway.

This commit does not solve the problem of failed/interrupted commands
leaving a hash around.