bin/memo: memo -c clears instead of refreshes
By duplicating functionality with memo -c, I made the semantics
consfusing and the UI annoying.

I often encountered this:

   memo -c <long-running-command>

When it was done I would return to the terminal and:

   1. I'd see the output
   2. Decide I need to munge it some more.
   3. Press <up> and add a filter.
   4. Lost my state because I still had -c in there, therefore I need to
      wait again.
   5. Silently curse at myself.

Fix this by making `memo -c` only clear. This immediate feedback makes
it obvious that I need to re-run without -c, and eliminates this
surprisingly common class of mistakes.