bin/memo: don't memoize when interrupted
Sadly, this has the side-effect of clearing the memoization if the
memo(1) part of the pipeline has already completed but the pipeline is
interrupted later. Looking at my own usage of memo(1), this appears to
be a fringe use case.

I initially also wanted to (provide a flag for) not saving the
memoization if the process invoked by memo(1) returned an error code,
but it was very tricky. Some code I used for debugging the problem
highlighted in the comments:

  nocache() {
    # $? represents the signal on all handlers except for ERR, where it
    # represents the process error code. Note to self: find a clever way
    # to deal with this.
    declare -ri ret=$?
    declare -ri sig=$((ret - 128))
    trap -p INT
    trap - INT ERR # Clear traps.
    local reason
    echo "sig: $sig" >&2
    if (( sig == -127 )) ; then
      reason='command failed'
    else
      reason="killed by $(kill -l $sig)"
    fi
    read -a st < /proc/self/stat
    echo "memo: $reason (I am ${st[0]} sig:${st[30]} sigblk:${st[31]} sigign:${st[32]}) not memoizing" >&2
    # echo "$(</proc/self/status)"
    readarray rows < /proc/self/status
    unsigmask <<< "${rows[@]}"
    # echo "${rows[@]}"
    # cat - < /proc/self/status
    # exit $ret
  }