vim: switch to LSP, remove syntastic and vim-go
Feels a lot cleaner. For a while I considered moving from Syntastic to
ALE due to the latter's asynchronous capabilities, but I never worked up
the X to do it. Now I'm switching instead to a pure LSP approach:

 - LSP's seem to be the future.
 - It provided me a chance to experiment with neovim 0.5.0 (pre-release)
   native LSP functionality.

Syntastic is replaced with the `efm-langserver`, which is actually a
shim langserver that calls out to normal linting/formatting tools. For
now I'm only configured the basics (shellcheck).

Currently only the following languages have servers configured:

 - C/C++: clangd
 - Go: gopls
 - Rust: rust-analyzer
 - Bash (and Bourne sh): efm-langserver + shellcheck

Notably, this setup is not yet 100% good, because it throws a lot of
errors of one of the components (like clangd) isn't installed properly.
I still need to figure this out better.