vim: disable mouse for vim and nvim
I'm not sure why I used to have this enabled. Perhaps it worked
differently with vanilla vim.

Neovim (very) recently switched to enabling mouse= by default (it used
to be disabled). This made me notice because the mouse started working,
and it made it impossible for me to select text in the status line. I
copy/paste from the status line a lot. I never use the mouse. So just
disable this.