vim: remove unused mappings: <leader>s and <leader>/
The <leader>s one is especially weird. It seems to create two vertical
splits of the current buffers. I don't think I ever used it.

This came up because one of my new mappings was taking a long time to
"execute". It turns out this was due to conflicting mappings (Vim waits
for `timeoutlen` if one mapping is a prefix of another).