vim/lsp: improve support for files outside current working directory
The `start` arg of find_first_parent was previously ignored. Not
ignoring it has the side effect of returning relative paths from
vim.fs.dirname, which some language servers don't like (I tried e.g.
gopls). So make sure the path is absolute (`path_absolute`).

The removal of `vim.fs.dirname()` from the helper function (now renamed
`find_up` is due to a future LSP for which we want to find a dir, which
don't want to take the parent of.