vim/lsp: ensure input is absolute, not output
This was biting me because I was wondering why vim.fs.find did not find
a ./lua folder that was upwards of my cwd. Turns out it doesn't search
upwards of ".". That may be a bug, but it's easy enough to work around.
This also avoids the need to absolutize the result of find_up, so it's
more efficient to boot.