vim/lsp: allow overriding multi-lang LSP server capabilities
If I don't do this, I get persistent annoying notifications that
"method X is not supported". Neovim is not really at fault, the server
did claim it could do this in the `server_capabilities`.

The correct fix would be to do a dynamic capability check, but this
doesn't work today for two reasons:

 1. The (internal) language server in question doesn't implement dynamic
    checks for the methods we're concerned with (higlighting,
    formatting, inlay hints, ...)
 2. Neovim prefers (see `_supports_method`) the static
    `server_capabilities` section. It will not do a dynamic check if
    present. If/when #1 gets fixed, we can consider making a
    change to Neovim that uses dynamic checks, at least if the `bufnr`
    option was set.