vim: fix trailing-space-removal-on-save
This got mangled when I did the bug single quotes (') to double quotes
(") replacement. The effect was that the cursor would move to the last
modified line on save, which was annoying. Not it works back like it
used to, as in https://github.com/aktau/dotfiles/commit/be75ffcb.