vim/lsp: migrate to vim.lsp.config and vim.lsp.enable
I was putting this off because I needed to find a way to "inhibit" other
LSP servers if some override LSP server was found.

After finding that `vim.lsp.Config.root_dir` can also be a function
which inhibits, I found a way to do it via monkey-patching:
https://github.com/neovim/neovim/issues/31552.

The big upside is that I can move this local-only setup to where the
inhibiting LSP is defined, cleaning up my open-source configuration.
This means my configuration doesn't need to be centralized anymore.

I like it better, despite the fact that it's less efficient in some
sense (I haven't noticed real slowdown).