vim: migrate from telescope.nvim to fzf-lua
The ability to open multiple items in splits is great. Neither telescope
nor snacks pickers allow this (though it would only require a tiny patch
for snacks[^1]).

fzf-lua also appears to be snappier, and supports the FZF syntax that I
k now and love. The only "downside" is that fzf needs to be installed
and in your $PATH, but that's table stakes for me.

[^1]: https://github.com/folke/snacks.nvim/issues/1270#issuecomment-2802829756